Add Open4::pfork4 for executing arbitrary Ruby code
Like popen4, but instead of executing command line, call arbitrary Ruby
code in a child process via a lambda.

We must use lambda for passing the code, because the block argument of
pfork4 is reserved for controlling the pipes. This is analogous to
popen4.

pfork4 propagates any exception raised in the child process to the
parent process if pfork4 is called with a block. This makes error
handling in the parent process more natural than with checking child
process exit status (which would destroy all exception information too).