Ignore errors when closing file-descriptors.
In particular, if one of these functions are called with closefds=true
shortly after a failed TCPSocket, the ObjectSpace iterator will find that
socket; but even called closed? on the socket will throw an error. Since
failing to close an fd is unlikely to be a serious issue, we just ignore
any errors.