Handle exception if pipe closed
If pfork4 is called without a block and the code running in the fork
raises an exception, there is a good change that the read end of the IO
pipe for communicating exception information to the parent process will
already be closed.  In that case, raise an exception so that the parent
process can at least see it, if the parent is monitoring the child's
stderr.